Made a property counting the effective satellites for a faction.
As we are now displaying the satellites on the faction board, it might be convinient to have a reusable function that returns the correct count. For the final scoring, Ivits' space stations were already counted. However, these were not displayed on the faction board. Both can now call the property on player data.